Image Inpainting software used to generate HDF5 dataset.
This version is *not* the one used to generate the first dataset. The
first dataset stored data into binary files as interleaved BGR, and
required manual assembly using Python's HDF5 bindings.

This version directly uses the HDF5 bindings and generates all image
sizes automatically: 64x64, 128x128 and 256x256. It also stores the
image as planar BGR instead of interleaved BGR, and thus the datasets
are 25000x3xHxW instead of the original 25000xHxWx3. Lastly, APIs
unavailable in Python are used to request allocation of the datasets on
a 16MB-aligned boundary, which makes for an easier time meory-maping the
file.