Done storage and I/O aspects.
The on-disk format is now settled.

Each file contains numbered sessions. If training is launched with a new
file, session #1 is created and initialized randomly. Otherwise, the
highest-numbered session is copied into a new session and training
resumes from the last state achieved by the old session.

A session has metadata (such as Unix time started, .tar.gz'ed source
code at launch, arguments at launch and the PRNG state used for
initialization), and two snapshots that we flip between atomically under
control of a single integer knob.

The snapshots contain the entire training state of the trainer,
including parameters, velocities, the current value of the PRNG and
more. Because the state is so large it can't be atomically updated, and
thus we risk on-disk corruption and/or partial updates if we kill
training at the wrong time. That's why two snapshots exist; One is the
"current" one, and one is the "next" one. Data to be snapshotted is
written whole to the "next" snapshot, flushed, a flip of the buffers is
attempted, then flushed. Either the flip succeeds or it does not, but in
either case the flag points to a fully valid state the training had
reached.