Working trainer at last.
The key component is batch normalization (BN). It seems BN everywhere
may not be as good as BN only between Fire modules. An idea to be tested
in BN *within* modules, after the compressor but before the expanders.

The BN inference-time parameter computation has yet to be implemented.

Also important are the policy for what to L1/L2 decay and what not, and
the L1/L2 penalties and learning rate.

One last thing remains to be done for the training loop: The
saveIfBestSoFar() feature.